requirements.txt
websocket-client
kafka-python-ng
python-dotenv
apache-flink
pandas
scikit-learn
numpy
requests
